+++ packages/chrome-devtools/src/utils/chrome/snapshot-plugin.ts
@@ -16,20 +16,22 @@ const chromeDevtoolsPlugin: () => FederationRuntimePlugin = function () {
     beforeLoadRemoteSnapshot({ options }) {
       const { nativeGlobal } = runtimeHelpers.global;
 
-      if (options.inBrowser) {
+      if (!options || options.inBrowser) {
         const realLocalStorage = nativeGlobal.localStorage || localStorage;
         const debugModuleInfoStr = realLocalStorage.getItem(
           MODULE_DEVTOOL_IDENTIFIER,
         );
         if (
           debugModuleInfoStr &&
-          !nativeGlobal.__INIT_VMOK_CHROME_DEVTOOL_PLUGIN__ &&
-          !options?.id?.endsWith(':local')
+          !nativeGlobal.__INIT_VMOK_CHROME_DEVTOOL_PLUGIN__
         ) {
           const chromeDevtoolSnapshot = JSON.parse(debugModuleInfoStr);
           if (chromeDevtoolSnapshot) {
             runtimeHelpers.global.addGlobalSnapshot(chromeDevtoolSnapshot);
             nativeGlobal.__INIT_VMOK_CHROME_DEVTOOL_PLUGIN__ = true;
+            console.warn(
+              '[Module Federation Devtools]: You are using the chrome devtool to proxy online module',
+            );
           }
         }
       }