+++ packages/enhanced/src/lib/container/runtime/FederationRuntimePlugin.ts
@@ -277,7 +277,7 @@ class FederationRuntimePlugin {
       };
     }
     if (this.options && !this.options?.name) {
-      // the instance may get the same one if the name is the same https://github.com/module-federation/universe/blob/main/packages/runtime/src/index.ts#L18
+      // the instance may get the same one if the name is the same https://github.com/module-federation/core/blob/main/packages/runtime/src/index.ts#L18
       this.options.name =
         compiler.options.output.uniqueName || `container_${Date.now()}`;
     }