+++ packages/native-federation-typescript/src/index.test.ts
@@ -75,10 +75,6 @@ describe('index', () => {
                 children: [
                   { name: 'archiveHandler.d.ts' },
                   { name: 'typeScriptCompiler.d.ts' },
-                  {
-                    name: 'writeBundle',
-                    children: [{ name: 'host.d.ts' }, { name: 'remote.d.ts' }],
-                  },
                 ],
               },
             ],
@@ -241,13 +237,6 @@ describe('index', () => {
                     children: [
                       { name: 'archiveHandler.d.ts' },
                       { name: 'typeScriptCompiler.d.ts' },
-                      {
-                        name: 'writeBundle',
-                        children: [
-                          { name: 'host.d.ts' },
-                          { name: 'remote.d.ts' },
-                        ],
-                      },
                     ],
                   },
                 ],