+++ packages/nextjs-mf/src/plugins/NextFederationPlugin/index.ts
@@ -6,7 +6,6 @@
 'use strict';
 
 import type {
-  ModuleFederationPluginOptions,
   NextFederationPluginExtraOptions,
   NextFederationPluginOptions,
 } from '@module-federation/utilities';
@@ -147,7 +146,7 @@ export class NextFederationPlugin {
   private getNormalFederationPluginOptions(
     compiler: Compiler,
     isServer: boolean,
-  ): ModuleFederationPluginOptions {
+  ): moduleFederationPlugin.ModuleFederationPluginOptions {
     const defaultShared = this._extraOptions.skipSharingNextInternals
       ? {}
       : retrieveDefaultShared(isServer);
@@ -187,6 +186,9 @@ export class NextFederationPlugin {
         ...defaultShared,
         ...this._options.shared,
       },
+      ...(isServer
+        ? { manifest: { filePath: '' } }
+        : { manifest: { filePath: '/static/chunks' } }),
       // nextjs project needs to add config.watchOptions = ['**/node_modules/**', '**/@mf-types/**'] to prevent loop types update
       dts: this._options.dts ?? false,
     };