+++ packages/nextjs-mf/src/plugins/NextFederationPlugin/next-fragments.ts
@@ -59,9 +59,11 @@ export const applyPathFixes = (compiler: Compiler, options: any) => {
       //@ts-ignore
       rule.oneOf.forEach((oneOfRule) => {
         if (hasLoader(oneOfRule, 'react-refresh-utils')) {
-          oneOfRule.exclude = [oneOfRule.exclude, /universe\/packages/].filter(
-            (i) => i,
-          );
+          oneOfRule.exclude = [
+            oneOfRule.exclude,
+            /universe\/packages/,
+            /core\/packages/,
+          ].filter((i) => i);
         }
       });
     }