+++ packages/chrome-devtools/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @module-federation/devtools
 
+## 0.1.12
+
+### Patch Changes
+
+- 2d761ae: fix: fix publish private error
+  - @module-federation/sdk@0.1.12
+
+## 0.1.11
+
+### Patch Changes
+
+- 99b23f7: fix(chrome-devtools): fix options undefined cause error
+- f8c6db7: chore(devtools): add modern lib to publish npm package
+  - @module-federation/sdk@0.1.11
+
+## 0.0.26
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.10
+
 ## 0.0.25
 
 ### Patch Changes