+++ packages/managers/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @module-federation/managers
 
+## 0.1.12
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.12
+
+## 0.1.11
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.11
+
+## 0.1.10
+
+### Patch Changes
+
+- @module-federation/sdk@0.1.10
+
 ## 0.1.9
 
 ### Patch Changes